tripbuilder/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.2.0
WTForms>=3.1.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
werkzeug>=2.3.0
boto3>=1.28.0
reportlab>=4.0.0
Pillow>=10.0.0